+++ src/Importers/HidroPretro10/Injectors/ClavesHyp.php
@@ -1,0 +1,48 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpCfdi\SatCatalogosPopulate\Importers\HidroPretro10\Injectors;
+
+use PhpCfdi\SatCatalogosPopulate\AbstractCsvInjector;
+use PhpCfdi\SatCatalogosPopulate\Database\DataFields;
+use PhpCfdi\SatCatalogosPopulate\Database\DataTable;
+use PhpCfdi\SatCatalogosPopulate\Database\DateDataField;
+use PhpCfdi\SatCatalogosPopulate\Database\TextDataField;
+use PhpCfdi\SatCatalogosPopulate\Utils\CsvFile;
+use RuntimeException;
+
+class ClavesHyp extends AbstractCsvInjector
+{
+    public function checkHeaders(CsvFile $csv): void
+    {
+        $csv->move(3);
+        $expected = [
+            'Clave producto',
+            'Descripción ',
+            'Clave Unidad',
+            'SubProducto',
+            'Fecha inicio de vigencia',
+            'Fecha fin de vigencia',
+        ];
+        $headers = $csv->readLine();
+
+        if ($expected !== $headers) {
+            throw new RuntimeException("The headers did not match on file {$this->sourceFile()}");
+        }
+
+        $csv->next();
+    }
+
+    public function dataTable(): DataTable
+    {
+        return new DataTable('hyp_10_claves_hyp', new DataFields([
+            new TextDataField('id'),
+            new TextDataField('descripcion'),
+            new TextDataField('clave_unidad'),
+            new TextDataField('subproductos'),
+            new DateDataField('vigencia_desde'),
+            new DateDataField('vigencia_hasta'),
+        ]));
+    }
+}