+++ tests/Unit/Importers/HidroPetro10/Injectors/SubproductosHypTest.php
@@ -1,0 +1,71 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpCfdi\SatCatalogosPopulate\Tests\Unit\Importers\HidroPetro10\Injectors;
+
+use PhpCfdi\SatCatalogosPopulate\AbstractCsvInjector;
+use PhpCfdi\SatCatalogosPopulate\Database\DateDataField;
+use PhpCfdi\SatCatalogosPopulate\Database\TextDataField;
+use PhpCfdi\SatCatalogosPopulate\Database\UsePreviousIfEmptyDataField;
+use PhpCfdi\SatCatalogosPopulate\Importers\HidroPretro10\Injectors\SubproductosHyp;
+use PhpCfdi\SatCatalogosPopulate\InjectorInterface;
+use PhpCfdi\SatCatalogosPopulate\Tests\TestCase;
+use PhpCfdi\SatCatalogosPopulate\Utils\ArrayProcessors\RightTrim;
+use PhpCfdi\SatCatalogosPopulate\Utils\CsvFile;
+use RuntimeException;
+
+final class SubproductosHypTest extends TestCase
+{
+    private string $sourceFile;
+
+    private SubproductosHyp $injector;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+        $this->sourceFile = $this->utilFilePath('hidropetro10/c_SubProductoHYP.csv');
+        $this->injector = new SubproductosHyp($this->sourceFile);
+    }
+
+    public function testClavesRetencionExtendsAbstractCsvInjector(): void
+    {
+        $this->assertInstanceOf(AbstractCsvInjector::class, $this->injector);
+        $this->assertInstanceOf(InjectorInterface::class, $this->injector);
+    }
+
+    public function testCheckHeadersOnValidSource(): void
+    {
+        $csv = new CsvFile($this->sourceFile, new RightTrim());
+        $this->injector->checkHeaders($csv);
+
+        $this->assertSame(4, $csv->position(), 'The csv position is on the first content line');
+    }
+
+    public function testCheckHeadersOnInvalidSource(): void
+    {
+        $csv = new CsvFile($this->utilFilePath('sample.csv'));
+
+        $this->expectException(RuntimeException::class);
+        $this->expectExceptionMessage('The headers did not match on file');
+        $this->injector->checkHeaders($csv);
+    }
+
+    public function testDataTable(): void
+    {
+        $dataTable = $this->injector->dataTable();
+        $this->assertSame('hyp_10_subproductos_hyp', $dataTable->name());
+        $expectedClasses = [
+            'id' => TextDataField::class,
+            'descripcion' => TextDataField::class,
+            'categoria' => UsePreviousIfEmptyDataField::class,
+            'vigencia_desde' => DateDataField::class,
+            'vigencia_hasta' => DateDataField::class,
+        ];
+        $this->assertSame(array_keys($expectedClasses), $dataTable->fields()->keys());
+        foreach ($expectedClasses as $key => $classname) {
+            $this->assertInstanceOf($classname, $dataTable->fields()->get($key));
+        }
+        $this->assertSame(['id'], $dataTable->primaryKey());
+    }
+}